Move const eval machine into its own module
